updated tests

+++ pychron/core/helpers/logger_setup.py
@@ -16,10 +16,11 @@
 
 # =============enthought library imports=======================
 # =============standard library imports ========================
-import os
 import logging
-from logging.handlers import RotatingFileHandler
+import os
 import shutil
+from logging.handlers import RotatingFileHandler
+
 # =============local library imports  =========================
 from pychron.paths import paths
 from pychron.core.helpers.filetools import list_directory
@@ -90,7 +91,7 @@ def tail(f, lines=20):
 #         logger.addHandler(h)
 
 
-def logging_setup(name, use_archiver=True, root=None, **kw):
+def logging_setup(name, use_archiver=True, root=None, use_file=True, **kw):
     """
     """
     # set up deprecation warnings
@@ -113,29 +114,35 @@ def logging_setup(name, use_archiver=True, root=None, **kw):
                      root=bdir)
         a.clean()
 
-    # create a new logging file
-    logname = '{}.current.log'.format(name)
-    logpath = os.path.join(bdir, logname)
+    if use_file:
+        # create a new logging file
+        logname = '{}.current.log'.format(name)
+        logpath = os.path.join(bdir, logname)
 
-    if os.path.isfile(logpath):
-        backup_logpath, _cnt = unique_path2(bdir, name, delimiter='-', extension='.log', width=5)
+        if os.path.isfile(logpath):
+            backup_logpath, _cnt = unique_path2(bdir, name, delimiter='-', extension='.log', width=5)
 
-        shutil.copyfile(logpath, backup_logpath)
-        os.remove(logpath)
+            shutil.copyfile(logpath, backup_logpath)
+            os.remove(logpath)
 
-        ps = list_directory(bdir, filtername=logname, remove_extension=False)
-        for pi in ps:
-            _h, t = os.path.splitext(pi)
-            v = os.path.join(bdir, pi)
-            shutil.copyfile(v, '{}{}'.format(backup_logpath, t))
-            os.remove(v)
+            ps = list_directory(bdir, filtername=logname, remove_extension=False)
+            for pi in ps:
+                _h, t = os.path.splitext(pi)
+                v = os.path.join(bdir, pi)
+                shutil.copyfile(v, '{}{}'.format(backup_logpath, t))
+                os.remove(v)
 
     root = logging.getLogger()
     root.setLevel(gLEVEL)
     shandler = logging.StreamHandler()
-    rhandler = RotatingFileHandler(
-        logpath, maxBytes=1e7, backupCount=5)
-    for hi in (shandler, rhandler):
+
+    handlers = [shandler]
+    if use_file:
+        rhandler = RotatingFileHandler(
+                logpath, maxBytes=1e7, backupCount=5)
+        handlers.append(rhandler)
+
+    for hi in handlers:
         hi.setLevel(gLEVEL)
         hi.setFormatter(logging.Formatter(gFORMAT))
         root.addHandler(hi)
